+++ apps/automation-v2/src/common/discord.ts
@@ -1,4 +1,4 @@
-import { Logger } from '@mountainpath9/overlord';
+import { Logger, TaskContext, TaskException } from '@mountainpath9/overlord';
 import { WebhookClient, MessageCreateOptions } from 'discord.js';
 
 interface DiscordChannel {
@@ -30,6 +30,31 @@ export function decodeWebhookUrl(url: string): { id: string; token: string } {
     throw new Error('Unable to decode discord webhook url');
   }
   return { id: m[1], token: m[2] };
+}
+
+export async function discordNotifyTaskException(ctx: TaskContext, te: TaskException) {
+  const content = [
+    `**TEMPLE Task Failed**`,
+    `task label: ${te.label}`,
+    `task id: ${te.taskId}`,
+    `task phase: ${te.phase}`,
+  ];
+
+  if (te.exception instanceof Error) {
+    content.push(`exception type: Error`);
+    // We truncate the message here as discord doesn't like large contents
+    const message =
+      te.exception.message.length > 997
+        ? te.exception.message.substring(0, 997) + '...'
+        : te.exception.message;
+    content.push(`exception message: ${message}`);
+  } else {
+    content.push(`exception type: unknown`);
+  }
+
+  const webhookUrl = await ctx.getSecret(DISCORD_WEBHOOK_URL_KEY);
+  const discord = await connectDiscord(webhookUrl, ctx.logger);
+  await discord.postMessage({ content: content.join('\n') });
 }
 
 const DISCORD_URL_RE = new RegExp(
